Create data view for externally refreshed data
